🎨 Split ProgressBar
